+++ src/main/java/org/truffleruby/options/Options.java
@@ -67,8 +67,6 @@ public final class Options {
     public final boolean HOST_INTEROP;
     /** --trace-calls=true */
     public final boolean TRACE_CALLS;
-    /** --pattern-matching=false */
-    public final boolean PATTERN_MATCHING;
     /** --patching=true */
     public final boolean PATCHING;
     /** --hashing-deterministic=false */
@@ -230,7 +228,6 @@ public final class Options {
         POLYGLOT_STDIO = options.hasBeenSet(OptionsCatalog.POLYGLOT_STDIO_KEY) ? options.get(OptionsCatalog.POLYGLOT_STDIO_KEY) : EMBEDDED || !NATIVE_PLATFORM;
         HOST_INTEROP = env.isHostLookupAllowed() && (options.get(OptionsCatalog.HOST_INTEROP_KEY));
         TRACE_CALLS = options.get(OptionsCatalog.TRACE_CALLS_KEY);
-        PATTERN_MATCHING = options.get(OptionsCatalog.PATTERN_MATCHING_KEY);
         PATCHING = options.get(OptionsCatalog.PATCHING_KEY);
         HASHING_DETERMINISTIC = options.get(OptionsCatalog.HASHING_DETERMINISTIC_KEY);
         VIRTUAL_THREAD_FIBERS = options.get(OptionsCatalog.VIRTUAL_THREAD_FIBERS_KEY);
@@ -346,8 +343,6 @@ public final class Options {
                 return HOST_INTEROP;
             case "ruby.trace-calls":
                 return TRACE_CALLS;
-            case "ruby.pattern-matching":
-                return PATTERN_MATCHING;
             case "ruby.patching":
                 return PATCHING;
             case "ruby.hashing-deterministic":